Move initialization code to couchbase package
This package also needs initialization,
move here to avoid cyclic deps
